requirements.txt
annotated-types
anyio
certifi
charset-normalizer
click
docx
fastapi
h11
httptools
idna
lxml
packaging
pillow
psycopg2-binary
pydantic
pydantic_core
PyMuPDF
pytesseract
python-docx
python-dotenv
python-multipart
PyYAML
requests
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
faiss-cpu
numpy
sentence-transformers
openai
ultralytics